Added CreateSims
Sim creation methods from Simulation moved to CreateSims.
